Add service option allowing different SPN prefix
On Kerberos connections, evil-winrm does currently not specify a service
name explicitly. This will lead to the default SPN prefix 'HTTP', which
is fine in most cases. However, in certain scenarios it is useful to
allow different SPN names (e.g. like WSMAN).

This commit introduces a new command line argument (-S, --service),
which can be used to specify the SPN prefix explicitly. If not
specified, the default prefix (HTTP) is used.